backend/requirements.txt
absl-py==2.1.0
alabaster==0.7.16
anyio==4.4.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
black==24.8.0
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
contourpy==1.2.0
cryptography==42.0.2
curio==1.6
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.0
decorator==5.1.1
deepface==0.0.84
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
Django==5.0.6
docrepr==0.2.0
docutils==0.20.1
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Mail==0.9.1
flatbuffers==23.5.26
fonttools==4.48.1
fqdn==1.5.1
gast==0.5.4
gdown==5.1.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
imageio==2.35.0
imagesize==1.4.1
imgaug==0.4.0
iniconfig==2.0.0
ipykernel==6.27.1
ipyparallel==8.8.0
ipython==8.19.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
jwt==1.3.1
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
mask-rcnn==2.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
ml-dtypes==0.2.0
mtcnn==0.1.1
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.5.8
networkx==3.3
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.2.0
pyparsing==3.1.1
PySocks==1.7.1
pytest==8.1.1
pytest-asyncio==0.21.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.2
pyzmq==25.1.2
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.14
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
rsa==4.9
scikit-image==0.24.0
scipy==1.14.0
Send2Trash==1.8.3
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.4.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.0
stack-data==0.6.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-macos==2.15.0
termcolor==2.4.0
terminado==0.18.1
testpath==0.6.0
tifffile==2024.8.10
tinycss2==1.3.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
trio==0.26.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
urllib3==2.2.0
virtualenv==20.25.1
wcwidth==0.2.12
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
widgetsnbextension==4.0.11
wrapt==1.14.1
djangprestframwork
django-cors-headers
djangorestframework-simplejwt